+++ Services/Udalost/Udalost_Service/Repositories/UdalostRepository.cs
@@ -19,78 +19,82 @@ namespace Udalost_Service.Repositories
         {
             _connectionString = connectionString;
         }
-        public void AddMessage(string message)
+        public async Task AddMessage(string message)
         {
             //-------------Description: Deserializace Json objektu na základní typ zprávy
             var envelope = JsonConvert.DeserializeObject<EventBase>(message);
             //-------------Description: Rozhodnutí o typu získazné zprávy
             switch (envelope.MessageType)
             {
-                case EventEnum.MessageType.DochazkaCreate:
+                case MessageType.DochazkaCreate:
                     //-------------Description: Kontrola verze zprávy 
                     if (envelope.Version == 1)
                     {
                         //-------------Description: Deserializace zprávy do správného typu a odeslání k uložení do DB; 
-                        this.Add(JsonConvert.DeserializeObject<EventUdalostCreate>(message));
+                        await this.Add(JsonConvert.DeserializeObject<EventUdalostCreate>(message));
                     }
                     break;
-                case EventEnum.MessageType.DochazkaRemove:
+                case MessageType.DochazkaRemove:
                     if (envelope.Version == 1)
                     {
-                        this.Remove(JsonConvert.DeserializeObject<EventUdalostRemove>(message));
+                        await this.Remove(JsonConvert.DeserializeObject<EventUdalostRemove>(message));
                     }
                     break;
-               case EventEnum.MessageType.DochazkaUpdate:
+               case MessageType.DochazkaUpdate:
                     if (envelope.Version == 1)
                     {
-                        this.Update(JsonConvert.DeserializeObject<EventUdalostUpdate>(message));
+                        await this.Update(JsonConvert.DeserializeObject<EventUdalostUpdate>(message));
                     }
                     break;
-                default:
-                    break;
+      
             }
         }
-        public void Add(EventUdalostCreate msg)
+        public async Task Add(EventUdalostCreate msg)
         {
             using (var db = new UdalostFactory(_connectionString).CreateDbContext())
             {
                 var add = new Udalost()
                 {
                     DatumOd = msg.DatumOd,
-                    DatumDo = msg.DatumDo,                    
-                    Nazev = msg.Nazev,
+                    DatumDo = msg.DatumDo,        
+                    UdalostTypId = msg.UdalostTypId,
+                    Popis = msg.Popis,
                     UzivatelId = msg.UzivatelId,
                     DatumZadal = msg.DatumZadal,
                 };                
                 db.Add(add);
-                db.SaveChanges();
+                await db.SaveChangesAsync();
             }          
         }
-        public void Remove(EventUdalostRemove msg)
+        public async Task Remove(EventUdalostRemove msg)
         {
             using (var db = new UdalostFactory(_connectionString).CreateDbContext())
             {
                 var remove = db.Udalosti.FirstOrDefault(b => b.Id == msg.UdalostId);
-                db.Udalosti.Remove(remove);
-                db.SaveChanges();
+                if (remove != null) {
+                    db.Udalosti.Remove(remove);
+                    await db.SaveChangesAsync();                    
+                } 
             }
         }
-        public bool Update(EventUdalostUpdate msg)
+        public async Task Update(EventUdalostUpdate msg)
         {
             using (var db = new UdalostFactory(_connectionString).CreateDbContext())
             {
                 var forUpdate = db.Udalosti.FirstOrDefault(b => b.Id == msg.UdalostId);
                 forUpdate.DatumOd = msg.DatumOd;
                 forUpdate.DatumDo = msg.DatumDo;
-                forUpdate.Nazev = msg.Nazev;
+                forUpdate.UdalostTypId = msg.UdalostTypId;
+                forUpdate.Popis = msg.Popis;
                 forUpdate.UzivatelId = msg.UzivatelId;
                 forUpdate.DatumZadal = msg.DatumZadal;
                 db.Udalosti.Update(forUpdate);
-                db.SaveChanges();
-                return true;
+                await db.SaveChangesAsync();
+                
             }
 
         }
+
         //public Udalost Get(int id)
         //{
         //    using (var db = new UdalostFactory(_connectionString).CreateDbContext())
